버그 수정
체력바 위치 및 UI 고정
체력바 표기오류 수정
적이 플레이어 아래쪽으로 이동하는 버그 수정
bat가 플레이어와 닿았을때 공격하도록 수정
플레이어와 bat가 닿지 않아 공격하지 않는 버그 발생
=>적이 플레이어와의 거리 1에서 멈추기->플레이어와 닿으면 멈추기 로 수정

+++ Assets/Script/Game/HealthBar.cs
@@ -10,6 +10,6 @@ public class HealthBar : MonoBehaviour
 
     void Update()   
     {
-        Hp.fillAmount = GameManager.player_current_HP / GameManager.player_HP;
+        Hp.fillAmount = (float)GameManager.player_current_HP / GameManager.player_HP;
     }
 }
